# docs/openapi/waflo-openapi-current.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/public/m/{slug}": {
      "get": {
        "operationId": "PublicMenuController_getPublicMenu",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public active menu by business slug.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "business": {
                      "id": "bus_123",
                      "name": "Tavrix Cafe",
                      "slug": "tavrix-cafe",
                      "type": "cafe",
                      "logoUrl": null,
                      "coverUrl": null,
                      "currency": "IQD",
                      "language": "ar",
                      "city": "Baghdad",
                      "menuTemplateId": "waflo-warm",
                      "menuThemeOverrides": null
                    },
                    "categories": [
                      {
                        "id": "cat_123",
                        "nameAr": "Hot Drinks",
                        "nameEn": null,
                        "sortOrder": 0,
                        "items": [
                          {
                            "id": "item_123",
                            "nameAr": "Turkish Coffee",
                            "nameEn": null,
                            "descriptionAr": "Traditional strong coffee.",
                            "descriptionEn": null,
                            "price": "3000",
                            "imageUrl": null,
                            "isAvailable": true,
                            "sortOrder": 0
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "public menu"
        ]
      }
    },
    "/public/m/{slug}/items/{itemId}": {
      "get": {
        "operationId": "PublicMenuController_getPublicItem",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public available item by business slug."
          }
        },
        "tags": [
          "public menu"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tavrix Menu API",
    "description": "Tavrix Menu backend API.\n\nProduction and staging clients must send real Clerk JWTs with Authorization: Bearer <clerk-jwt>.\n\nDevelopment can use dev tokens only when NODE_ENV=development.\n\nDevelopment auth header example:\n\nAuthorization: Bearer dev:user_tavrix_owner;email=owner@tavrix.local;name=Tavrix%20Owner",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}